Scream specific update for end-to-end workflow (#2384)
This PR includes various changes required to work with the latest round
of scream native output with the end-to-end workflow (training and
prognostic_run diag).

Significant internal changes:
- `external/loaders/loaders/mappers/_nudged/_nudged.py`: renaming from
`dQxwind` to `dQu` - there's only one definition of horizontal wind in
scream and the output is `u` and `v`
- `projects/scream/scream_post/format_output.py`: adding additional
fields we need to post process for training surface radiative fluxes
- `projects/scream/scream_post/translate_prognostic_output.py`: we use
this script to post process scream online prognostic run output, then
run them through prognostic_run diag
- `projects/scream/scream_post/util.py`: moved several functions shared
between format_output and translate_prognostic_output
- `constants.py`, `derived_variables.py`, `diurnal_cycle.py`: scream
outputs `surf_evap`, which is different from fv3's `LHTFLsfc`. We allow
`surf_evap` to be a valid variable when present in the dataset and treat
it as evaporation.