Add transformer for normalizing individual spatial grid points (#2382)
Add a new encoder `_ScaleSpatialConcatZTransformer` built using
`build_scale_spatial_concat_z_transformer` to perform a per grid-point
normalization. This matches the normalization used by Acomano et al. in
their coupled atmosphere--ocean hybrid reservoir modeling.

Also adds a workaround for the batch processing in the training to
account for the fact that an encoder may expect data with or without a
halo. Previously we assumed that when a halo was provided in the data to
process and the output should be trimmed, the encoder would accept halo
data. We should probably standardize this.

Added public API:
- `build_scale_spatial_concat_z_transformer` - a function to create a
transformer object that normalizes per grid point and combines along the
z features. (Currently requires all z features to be equivalent across
included variables)

Refactored public API:
- `encode_columns` and `decode_columns` removed in favor of adding
`encode/decode_txyz` methods to the base `Transformer` class.

Significant internal changes:
- Added a workaround for transformer use in processing tfdataset batch
data that may or may not expect data with a halo

- [x] Tests added


Coverage reports (updated automatically):